cloud_functions/iss_api_query_assistant/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
functions-framework==3.*
google-cloud-secret-manager==2.16.4
requests==2.31.0
google-cloud-logging==3.9.0
